crawler/requirements.txt
scrapy==2.11.2
pymysql==1.1.1
python-dotenv==1.0.1

numpy<2
sentence-transformers==2.7.0
transformers==4.41.2

--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.2+cpu
